refactor: GroupService 에서 Join 서비스 분리
