+++ backend/rush/src/main/java/rush/rush/controller/group/JoinGroupController.java
@@ -9,18 +9,18 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 import rush.rush.security.CurrentUser;
 import rush.rush.security.user.UserPrincipal;
-import rush.rush.service.group.GroupService;
+import rush.rush.service.group.JoinGroupService;
 
 @RequiredArgsConstructor
 @RestController
 @RequestMapping("/groups")
 public class JoinGroupController {
 
-    private final GroupService groupService;
+    private final JoinGroupService joinGroupService;
 
     @PostMapping("/join")
     public ResponseEntity<Void> join(@RequestParam(value = "invitation_code") String invitationCode, @CurrentUser UserPrincipal userPrincipal) {
-        Long groupId = groupService.join(invitationCode, userPrincipal.getUser());
+        Long groupId = joinGroupService.join(invitationCode, userPrincipal.getUser());
 
         return ResponseEntity.created(URI.create("/api/groups/" + groupId))
                 .build();
